# elpe/ui/sampleResponseAndSwagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Risk analysis REST API",
        "version": "1.0.0",
        "title": "We metrix Risk analysis API",
        "contact": {
            "name": "John Garos, head of big data",
            "url": "https://wemetrix.com/",
            "email": "info@wemetrix.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8090",
    "basePath": "/",
...
    "paths": {
...
        "/scenarios/fetch": {
            "get": {
                "tags": [
                    "scenario-controller"
                ],
                "summary": "getScenarios",
                "operationId": "getScenariosUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "snapshotDate",
                        "in": "query",
                        "description": "snapshotDate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ScenarioContainerUiDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/scenarios/template": {
            "get": {
                "tags": [
                    "scenario-controller"
                ],
                "summary": "template",
                "operationId": "templateUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "authenticated",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "authorities[0].authority",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comparisonPeriod",
                        "in": "query",
                        "description": "comparisonPeriod",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "contextId",
                        "in": "query",
                        "description": "contextId",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "credentials",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "details",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "principal",
                        "in": "query",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "snapshotDate",
                        "in": "query",
                        "description": "snapshotDate",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ScenarioContainerUiDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
...
        "KeyValueDto": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            },
            "title": "KeyValueDto"
        },
        "KeyValueMultipleDto": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            },
            "title": "KeyValueMultipleDto"
        },
        "KeyValuesDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValueDto"
                    }
                }
            },
            "title": "KeyValuesDto"
        },
...
        "ScenarioContainerUiDto": {
            "type": "object",
            "properties": {
                "contextId": {
                    "type": "integer",
                    "format": "int32"
                },
                "scenarios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ScenarioUIDto"
                    }
                },
                "snapshotDate": {
                    "type": "string"
                }
            },
            "title": "ScenarioContainerUiDto"
        },
        "ScenarioCustomerParametersUIDto": {
            "type": "object",
            "properties": {
                "marginChange": {
                    "type": "string",
                    "description": "Change in profit margins"
                },
                "marginChangeCli": {
                    "type": "string",
                    "description": "Change in profit margins"
                },
                "maxDso": {
                    "type": "string",
                    "description": "Maximum accepted DSO (days)"
                },
                "maxDsoCli": {
                    "type": "string",
                    "description": "Maximum accepted DSO (days)"
                },
                "maxLimitGrowth": {
                    "type": "string",
                    "description": "Maximum limit growth"
                },
                "maxLimitGrowthCli": {
                    "type": "string",
                    "description": "Maximum limit growth"
                },
                "maxLimitReduction": {
                    "type": "string",
                    "description": "Maximum limit reduction"
                },
                "maxLimitReductionCli": {
                    "type": "string",
                    "description": "Maximum limit reduction"
                },
                "minAcceptedLimitUse": {
                    "type": "string",
                    "description": "Minimum accepted limit use"
                },
                "minAcceptedLimitUseCli": {
                    "type": "string",
                    "description": "Minimum accepted limit use"
                },
                "minAcceptedRwMargin": {
                    "type": "string",
                    "description": "Minimum accepted RW Margin"
                },
                "minAcceptedRwMarginCli": {
                    "type": "string",
                    "description": "Minimum accepted RW Margin"
                },
                "minLimitGrowth": {
                    "type": "string",
                    "description": "Minimum limit growth"
                },
                "minLimitGrowthCli": {
                    "type": "string",
                    "description": "Minimum limit growth"
                },
                "minLimitReduction": {
                    "type": "string",
                    "description": "Minimum limit reduction"
                },
                "minLimitReductionCli": {
                    "type": "string",
                    "description": "Minimum limit reduction"
                },
                "pdChange": {
                    "type": "string",
                    "description": "Change in default probabilities"
                },
                "pdChangeCli": {
                    "type": "string",
                    "description": "Change in default probabilities"
                },
                "salesChange": {
                    "type": "string",
                    "description": "Sales change"
                },
                "salesChangeCli": {
                    "type": "string",
                    "description": "Sales change"
                }
            },
            "title": "ScenarioCustomerParametersUIDto",
            "description": "Contain customer of scenario"
        },
        "ScenarioCustomerUIDto": {
            "type": "object",
            "properties": {
                "customerId": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "parameters": {
                    "$ref": "#/definitions/ScenarioCustomerParametersUIDto"
                }
            },
            "title": "ScenarioCustomerUIDto"
        },
...
        "ScenarioMainParametersUIDto": {
            "type": "object",
            "properties": {
                "creditAmount": {
                    "type": "string",
                    "description": "Target total credit limit"
                },
                "marginChange": {
                    "type": "string",
                    "description": "Change in profit margins"
                },
                "maxDso": {
                    "type": "string",
                    "description": "Maximum accepted DSO (days)"
                },
                "maxLimitGrowth": {
                    "type": "string",
                    "description": "Maximum limit growth"
                },
                "maxLimitReduction": {
                    "type": "string",
                    "description": "Maximum limit reduction"
                },
                "minAcceptedLimitUse": {
                    "type": "string",
                    "description": "Minimum accepted limit use"
                },
                "minAcceptedRwMargin": {
                    "type": "string",
                    "description": "Minimum acceptable Risk Weighted Margin"
                },
                "minLimit": {
                    "type": "string",
                    "description": "Minimum acceptable limit amount"
                },
                "minLimitGrowth": {
                    "type": "string",
                    "description": "Minimum limit growth"
                },
                "minLimitPct": {
                    "type": "string",
                    "description": "Minimum accepted limit over annual turnover"
                },
                "minLimitReduction": {
                    "type": "string",
                    "description": "Minimum limit reduction"
                },
                "pdChange": {
                    "type": "string",
                    "description": "Change in default probabilities"
                },
                "raroc": {
                    "type": "string",
                    "description": "Intended Return on Equity"
                },
                "salesChange": {
                    "type": "string",
                    "description": "Sales change"
                },
                "significantDigits": {
                    "type": "string",
                    "description": "Number of limit's significant digits"
                },
                "userComments": {
                    "type": "string",
                    "description": "userComments"
                },
                "wacc": {
                    "type": "string",
                    "description": "WACC (Weighted Average Cost of Capital)"
                },
                "worstAcceptableScore": {
                    "type": "string",
                    "description": "Worst acceptable CAS"
                }
            },
            "title": "ScenarioMainParametersUIDto",
            "description": "Contain main parameters of scenario"
        },
        "ScenarioMainStatisticsUIDto": {
            "type": "object",
            "properties": {
                "allowances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValuesDto"
                    }
                },
                "limits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValuesDto"
                    }
                },
                "margins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValuesDto"
                    }
                },
                "scenarioStatistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValueMultipleDto"
                    }
                }
            },
            "title": "ScenarioMainStatisticsUIDto"
        },
        "ScenarioPortfolioParametersUIDto": {
            "type": "object",
            "properties": {
                "marginChange": {
                    "type": "string",
                    "description": "Change in profit margins"
                },
                "marginChangeCli": {
                    "type": "string",
                    "description": "Change in profit margins"
                },
                "maxDso": {
                    "type": "string",
                    "description": "Maximum accepted DSO (days)"
                },
                "maxDsoCli": {
                    "type": "string",
                    "description": "Maximum accepted DSO (days)"
                },
                "maxLimitGrowth": {
                    "type": "string",
                    "description": "Maximum limit growth"
                },
                "maxLimitGrowthCli": {
                    "type": "string",
                    "description": "Maximum limit growth"
                },
                "maxLimitReduction": {
                    "type": "string",
                    "description": "Maximum limit reduction"
                },
                "maxLimitReductionCli": {
                    "type": "string",
                    "description": "Maximum limit reduction"
                },
                "minAcceptedLimitUse": {
                    "type": "string",
                    "description": "Minimum accepted limit use"
                },
                "minAcceptedLimitUseCli": {
                    "type": "string",
                    "description": "Minimum accepted limit use"
                },
                "minAcceptedRwMargin": {
                    "type": "string",
                    "description": "Minimum acceptable Risk Weighted Margin"
                },
                "minAcceptedRwMarginCli": {
                    "type": "string",
                    "description": "Minimum acceptable Risk Weighted Margin"
                },
                "minLimitGrowth": {
                    "type": "string",
                    "description": "Minimum limit growth"
                },
                "minLimitGrowthCli": {
                    "type": "string",
                    "description": "Minimum limit growth"
                },
                "minLimitReduction": {
                    "type": "string",
                    "description": "Minimum limit reduction"
                },
                "minLimitReductionCli": {
                    "type": "string",
                    "description": "Minimum limit reduction"
                },
                "pdChange": {
                    "type": "string",
                    "description": "Change in default probabilities"
                },
                "pdChangeCli": {
                    "type": "string",
                    "description": "Change in default probabilities"
                },
                "salesChange": {
                    "type": "string",
                    "description": "Sales change"
                },
                "salesChangeCli": {
                    "type": "string",
                    "description": "Sales change"
                },
                "userComments": {
                    "type": "string",
                    "description": "userComments"
                },
                "userCommentsCli": {
                    "type": "string",
                    "description": "userComments"
                },
                "worstAcceptableScore": {
                    "type": "string",
                    "description": "Worst acceptable CAS"
                },
                "worstAcceptableScoreCli": {
                    "type": "string",
                    "description": "Worst acceptable CAS"
                }
            },
            "title": "ScenarioPortfolioParametersUIDto",
            "description": "Contain portfolio parameters of scenario"
        },
        "ScenarioPortfolioStatisticsUIDto": {
            "type": "object",
            "properties": {
                "allowances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValuesDto"
                    }
                },
                "limits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValuesDto"
                    }
                },
                "margins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValuesDto"
                    }
                },
                "portfolioStatistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValueMultipleDto"
                    }
                }
            },
            "title": "ScenarioPortfolioStatisticsUIDto"
        },
        "ScenarioPortfolioUIDto": {
            "type": "object",
            "properties": {
                "parameters": {
                    "$ref": "#/definitions/ScenarioPortfolioParametersUIDto"
                },
                "portfolioId": {
                    "type": "string",
                    "description": "portfolio id"
                },
                "portfolioName": {
                    "type": "string",
                    "description": "portfolio name"
                },
                "statistics": {
                    "$ref": "#/definitions/ScenarioPortfolioStatisticsUIDto"
                }
            },
            "title": "ScenarioPortfolioUIDto",
            "description": "Contain portfolio parameters and statistics of scenario"
        },
        "ScenarioProspectParametersUIDto": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "string",
                    "description": "Comments"
                },
                "customersNum": {
                    "type": "string",
                    "description": "Number of prospective customers"
                },
                "description": {
                    "type": "string",
                    "description": "Prospect description"
                },
                "grade": {
                    "type": "string",
                    "description": "Prospective customers' risk level"
                },
                "gradeInx": {
                    "type": "string",
                    "description": "Prospective customers' risk level"
                },
                "maxDso": {
                    "type": "string",
                    "description": "Maximum accepted DSO (days)"
                },
                "minAcceptedRwMargin": {
                    "type": "string",
                    "description": "Minimum accepted Risk Weighted Margin"
                },
                "mitigants": {
                    "type": "array",
                    "description": "Mitigants requested",
                    "items": {
                        "type": "string"
                    }
                },
                "mitigantsValues": {
                    "type": "array",
                    "description": "Mitigants requested",
                    "items": {
                        "type": "string"
                    }
                },
                "profitMargin": {
                    "type": "string",
                    "description": "Profit margin"
                },
                "prospectId": {
                    "type": "string",
                    "description": "The id of prospect"
                },
                "purchases": {
                    "type": "string",
                    "description": "Expected annual turnover per customer"
                }
            },
            "title": "ScenarioProspectParametersUIDto",
            "description": "Contain prospect parameters of scenario"
        },
        "ScenarioProspectStatisticsUIDto": {
            "type": "object",
            "properties": {
                "prospectStatistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValueMultipleDto"
                    }
                }
            },
            "title": "ScenarioProspectStatisticsUIDto"
        },
        "ScenarioProspectUIDto": {
            "type": "object",
            "properties": {
                "parameters": {
                    "$ref": "#/definitions/ScenarioProspectParametersUIDto"
                },
                "statistics": {
                    "$ref": "#/definitions/ScenarioProspectStatisticsUIDto"
                }
            },
            "title": "ScenarioProspectUIDto"
        },
        "ScenarioUIDto": {
            "type": "object",
            "properties": {
                "customers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ScenarioCustomerUIDto"
                    }
                },
                "deletedCustomers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "deletedProspects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "editable": {
                    "type": "boolean"
                },
                "parameters": {
                    "$ref": "#/definitions/ScenarioMainParametersUIDto"
                },
                "portfolios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ScenarioPortfolioUIDto"
                    }
                },
                "prospects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ScenarioProspectUIDto"
                    }
                },
                "scenarioId": {
                    "type": "integer",
                    "format": "int32",
                    "description": "Scenario id"
                },
                "scenarioName": {
                    "type": "string",
                    "description": "Scenario name"
                },
                "scenarioOwnerName": {
                    "type": "string",
                    "description": "Scenario owner name"
                },
                "scenarioOwnerlogin": {
                    "type": "string",
                    "description": "Scenario owner login name"
                },
                "statistics": {
                    "$ref": "#/definitions/ScenarioMainStatisticsUIDto"
                }
            },
            "title": "ScenarioUIDto",
            "description": "Scenario container"
        },
...
    }
}